Restructures updateSchema to extract transforms.
