+++ src/index.js
@@ -1,3 +1,4 @@
 export cashayReducer from './normalize/duck';
 export Cashay from './Cashay';
 export HTTPTransport from './HTTPTransport';
+export transformSchema from './schema/transformSchema';
